fix: move arg0 handling out of codex-linux-sandbox and into its own crate
